Alter MySQL schema accordingly

+++ system/ee/installer/schema/mysql_schema.php
@@ -221,9 +221,9 @@ class EE_Schema
 			user_agent varchar(120) NOT NULL,
 			login_state varchar(32) NULL DEFAULT NULL,
 			fingerprint varchar(40) NOT NULL,
-			sess_start int(10) unsigned DEFAULT '0' NOT NULL,
-			auth_timeout int(10) unsigned DEFAULT '0' NOT NULL,
-			last_activity int(10) unsigned DEFAULT '0' NOT NULL,
+			sess_start bigint(10) unsigned DEFAULT '0' NOT NULL,
+			auth_timeout bigint(10) unsigned DEFAULT '0' NOT NULL,
+			last_activity bigint(10) unsigned DEFAULT '0' NOT NULL,
 			can_debug char(1) NOT NULL DEFAULT 'n',
             mfa_flag ENUM('skip','show','required') NOT NULL DEFAULT 'skip',
             pro_banner_seen char(1) NOT NULL DEFAULT 'n',
@@ -237,7 +237,7 @@ class EE_Schema
         $Q[] = "CREATE TABLE exp_throttle (
 			throttle_id int(10) unsigned NOT NULL auto_increment,
 			ip_address varchar(45) default '0' NOT NULL,
-			last_activity int(10) unsigned DEFAULT '0' NOT NULL,
+			last_activity bigint(10) unsigned DEFAULT '0' NOT NULL,
 			hits int(10) unsigned NOT NULL,
 			locked_out char(1) NOT NULL default 'n',
 			PRIMARY KEY `throttle_id` (`throttle_id`),
@@ -263,7 +263,7 @@ class EE_Schema
 			last_visitor_date bigint(10) unsigned default '0' NOT NULL,
 			most_visitors mediumint(7) NOT NULL default '0',
 			most_visitor_date bigint(10) unsigned default '0' NOT NULL,
-			last_cache_clear int(10) unsigned default '0' NOT NULL,
+			last_cache_clear bigint(10) unsigned default '0' NOT NULL,
 			PRIMARY KEY `stat_id` (`stat_id`),
 			KEY `site_id` (`site_id`)
 		)";
@@ -277,7 +277,7 @@ class EE_Schema
 			in_forum char(1) NOT NULL default 'n',
 			name varchar(" . USERNAME_MAX_LENGTH . ") default '0' NOT NULL,
 			ip_address varchar(45) default '0' NOT NULL,
-			date int(10) unsigned default '0' NOT NULL,
+			date bigint(10) unsigned default '0' NOT NULL,
 			anon char(1) NOT NULL,
 			PRIMARY KEY `online_id` (`online_id`),
 			KEY `date` (`date`),
@@ -324,7 +324,7 @@ class EE_Schema
 
         $Q[] = "CREATE TABLE exp_security_hashes (
 			hash_id int(10) unsigned NOT NULL auto_increment,
-			date int(10) unsigned NOT NULL,
+			date bigint(10) unsigned NOT NULL,
 			session_id varchar(40) default '0' NOT NULL,
 			hash varchar(40) NOT NULL,
 			PRIMARY KEY `hash_id` (`hash_id`),
@@ -335,7 +335,7 @@ class EE_Schema
 
         $Q[] = "CREATE TABLE exp_captcha (
 			captcha_id bigint(13) unsigned NOT NULL auto_increment,
-			date int(10) unsigned NOT NULL,
+			date bigint(10) unsigned NOT NULL,
 			ip_address varchar(45) default '0' NOT NULL,
 			word varchar(20) NOT NULL,
 			PRIMARY KEY `captcha_id` (`captcha_id`),
@@ -367,7 +367,7 @@ class EE_Schema
 			reset_id int(10) unsigned NOT NULL auto_increment,
 			member_id int(10) unsigned NOT NULL,
 			resetcode varchar(12) NOT NULL,
-			date int(10) NOT NULL,
+			date bigint(10) NOT NULL,
 			PRIMARY KEY `reset_id` (`reset_id`)
 		)";
 
@@ -460,12 +460,12 @@ class EE_Schema
 			ignore_list text NULL,
 			private_messages int(4) unsigned DEFAULT '0' NOT NULL,
 			accept_messages char(1) NOT NULL default 'y',
-			last_view_bulletins int(10) NOT NULL default 0,
+			last_view_bulletins bigint(10) NOT NULL default 0,
 			last_bulletin_date bigint(10) NOT NULL default 0,
 			ip_address varchar(45) default '0' NOT NULL,
 			join_date bigint(10) unsigned default '0' NOT NULL,
-			last_visit int(10) unsigned default '0' NOT NULL,
-			last_activity int(10) unsigned default '0' NOT NULL,
+			last_visit bigint(10) unsigned default '0' NOT NULL,
+			last_activity bigint(10) unsigned default '0' NOT NULL,
 			total_entries mediumint(8) unsigned NOT NULL default '0',
 			total_comments mediumint(8) unsigned NOT NULL default '0',
 			total_forum_topics mediumint(8) default '0' NOT NULL,
@@ -1305,7 +1305,7 @@ class EE_Schema
 			recipient_id int(10) unsigned NOT NULL default 0,
 			message_received char(1) NOT NULL default 'n',
 			message_read char(1) NOT NULL default 'n',
-			message_time_read int(10) unsigned NOT NULL default 0,
+			message_time_read bigint(10) unsigned NOT NULL default 0,
 			attachment_downloaded char(1) NOT NULL default 'n',
 			message_folder int(10) unsigned NOT NULL default 1,
 			message_authcode varchar(10) NOT NULL default '',
@@ -1393,7 +1393,7 @@ class EE_Schema
 			`bulletin_group` int(8) unsigned NOT NULL,
 			`bulletin_date` bigint(10) unsigned NOT NULL,
 			`hash` varchar(10) NOT NULL DEFAULT '',
-			`bulletin_expires` int(10) unsigned NOT NULL DEFAULT 0,
+			`bulletin_expires` bigint(10) unsigned NOT NULL DEFAULT 0,
 			`bulletin_message` text NOT NULL,
 			PRIMARY KEY `bulletin_id` (`bulletin_id`),
 			KEY `sender_id` (`sender_id`),
@@ -1548,7 +1548,7 @@ class EE_Schema
         // Developer log table
         $Q[] = "CREATE TABLE `exp_developer_log` (
 			`log_id` int(10) unsigned NOT NULL AUTO_INCREMENT,
-			`timestamp` int(10) unsigned NOT NULL,
+			`timestamp` bigint(10) unsigned NOT NULL,
 			`viewed` char(1) NOT NULL DEFAULT 'n',
 			`description` text NULL,
 			`function` varchar(100) NULL,
@@ -1574,8 +1574,8 @@ class EE_Schema
 			`user_agent` varchar(120) DEFAULT '',
 			`admin_sess` tinyint(1) DEFAULT '0',
 			`site_id` int(4) DEFAULT '1',
-			`expiration` int(10) DEFAULT '0',
-			`last_refresh` int(10) DEFAULT '0',
+			`expiration` bigint(10) DEFAULT '0',
+			`last_refresh` bigint(10) DEFAULT '0',
 			PRIMARY KEY (`remember_me_id`),
 			KEY `member_id` (`member_id`)
 		)";
@@ -1660,8 +1660,8 @@ class EE_Schema
 			`cookie_id` int(10) unsigned NOT NULL AUTO_INCREMENT,
 			`cookie_provider` varchar(50) NOT NULL,
 			`cookie_name` varchar(50) NOT NULL,
-			`cookie_lifetime` int(10) unsigned DEFAULT NULL,
-			`cookie_enforced_lifetime` int(10) unsigned DEFAULT NULL,
+			`cookie_lifetime` bigint(10) unsigned DEFAULT NULL,
+			`cookie_enforced_lifetime` bigint(10) unsigned DEFAULT NULL,
 			`cookie_title` varchar(200) NOT NULL,
 			`cookie_description` text NULL,
 			PRIMARY KEY (`cookie_id`)
